Create lower case column name in encode_categorical_variables()
JIRA:MADLIB-1202
The previous madlib.encode_categorical_variables() function generates
column name with some capital characters, including:
1. when you specify top_values, there will be a column name with suffix __MISC__
2. when you set encode_nulls as True, there will be a column name with suffix
__NULL
3. when the original column is boolean type, there will be column names
with suffix _True and _False

In the above cases, users have to use double quoting to query, which is
not conveninet.

This commit adresses this, and all of the three scenarios will generate
coloumn name with lower cases.