CNN model merging
This is a POC, to see if model merging works for CNN models learnt
using Keras on GPDB with multiple clusters.
The script at the moment assumes the input is CIFAR, and the images are
of 32x32x3 dimension. The architecture of the CNN is also hard-coded in
the code. This code should work with any input that has mini-batched its
input data, and follows the 32x32x3 input image size.

Co-authored-by: Rahul Iyer <riyer@apache.org>
Co-authored-by: Omer Arap <oarap@pivotal.io>