ModelVer: Add functions to convert bytea to bytea8 and double precision array
In order for the model versioned table to store all types of models
inclusing double precision arrays and bytea8 models, we decided to
store the model in a bytea column. This meant that we had to create the
following functions
1. Convert bytea to madlib.bytea8.
2. Create a C function to convert bytea to double precision array.

Here is how these functions will be used in the ModelVersion classes:
1. While storing the model, convert it into bytea (either from double
precision[] or bytea8)
2. While reading the model, convert bytea to either double precision[]
or bytea8.

Co-authored-by: Nikhil Kak <nkak@pivotal.io>