ModelVer: View Creation sql code for constructing latest_model view of DT
This creates the "latest_model" view for a repo containing models of
type decision tree.  (Does not include output_table_summary yet.)
Eventually, we want this to be a utility function that takes parameters
and works more generally.

Add quotes around "grouping" column.

None of the other columns require quotes.  But for some reason, if
we don't add them around "grouping", then the table validation code
on the repo table fails.  Even though when you list the columns using
\dt it does not show the quotes, and they were not there when it was
created.  For now an unsolved mystery; this will work, but we may
want to just automatically strip off the quotes for all columns in
case this happens again for another column.

Co-authored-by: Nandish Jayaram <njayaram@apache.org>