ModelVer + DT: Create separate cols for certain params in summary table
This makes the necessary changes in ModelVer to create new columns in
summary repo table. This also changes DT code to ensure the dev check
goes through fine. The test case for get_var_importance() depends on
code in random_forest.py_in, so some changes had to be made in that file
too to get DT dev check to green.